cleanup dead code + linting
